New structure for the BookingPage and components
