+++ src/Pages/BookingPage/components/AmountTheaterSeats.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 
 
-export default function SeatsTheater() {
+export default function AmountSeatsTheater() {
   //Creating useState method to give user a ticket counter.
   const [countAdult, setCountAdult] = useState<number>(1); //Always a start with 1
   const [countSenior, setCountSenior] = useState<number>(1);
